Replace old internal multipart, move bitcomet to transdroid
