+++ app/src/main/java/org/transdroid/daemon/DLinkRouterBT/DLinkRouterBTAdapter.java
@@ -1,25 +1,25 @@
 /*
  *	This file is part of Transdroid <http://www.transdroid.org>
- *	
+ *
  *	Transdroid is free software: you can redistribute it and/or modify
  *	it under the terms of the GNU General Public License as published by
  *	the Free Software Foundation, either version 3 of the License, or
  *	(at your option) any later version.
- *	
+ *
  *	Transdroid is distributed in the hope that it will be useful,
  *	but WITHOUT ANY WARRANTY; without even the implied warranty of
  *	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  *	GNU General Public License for more details.
- *	
+ *
  *	You should have received a copy of the GNU General Public License
  *	along with Transdroid.  If not, see <http://www.gnu.org/licenses/>.
- *	
+ *
  */
 package org.transdroid.daemon.DLinkRouterBT;
 
-import com.android.internalcopy.http.multipart.FilePart;
-import com.android.internalcopy.http.multipart.MultipartEntity;
-import com.android.internalcopy.http.multipart.Part;
+import com.android.internal.http.multipart.FilePart;
+import com.android.internal.http.multipart.MultipartEntity;
+import com.android.internal.http.multipart.Part;
 
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpResponse;
@@ -358,23 +358,23 @@ public class DLinkRouterBTAdapter implements IDaemonAdapter {
 
 			// @formatter:off
 			Torrent new_t = new Torrent(
-				i, 
-				tor.getString(BT_HASH), 
-				tor.getString(BT_CAPTION), 
+				i,
+				tor.getString(BT_HASH),
+				tor.getString(BT_CAPTION),
 				status,
 				null, // Not supported?
-				tor.getInt(BT_DOWNLOAD_RATE), 
-				tor.getInt(BT_UPLOAD_RATE), 
-				tor.getInt(BT_PEERS_CONNECTED), 
-				tor.getInt(BT_PEERS_TOTAL), 
-				tor.getInt(BT_SEEDS_CONNECTED), 
+				tor.getInt(BT_DOWNLOAD_RATE),
+				tor.getInt(BT_UPLOAD_RATE),
+				tor.getInt(BT_PEERS_CONNECTED),
+				tor.getInt(BT_PEERS_TOTAL),
+				tor.getInt(BT_SEEDS_CONNECTED),
 				tor.getInt(BT_SEEDS_TOTAL),
 				eta,
-				tor.getLong(BT_DONE), 
-				tor.getLong(BT_PAYLOAD_UPLOAD), 
-				tor.getLong(BT_SIZE), 
+				tor.getLong(BT_DONE),
+				tor.getLong(BT_PAYLOAD_UPLOAD),
+				tor.getLong(BT_SIZE),
 				tor.getLong(BT_DONE) / (float) tor.getLong(BT_SIZE),
-				Float.parseFloat(tor.getString(BT_COPYS)), 
+				Float.parseFloat(tor.getString(BT_COPYS)),
 				null,
 				null,
 				null,
@@ -402,11 +402,11 @@ public class DLinkRouterBTAdapter implements IDaemonAdapter {
 				// @formatter:off
 				torrentfiles.add(new TorrentFile(
 					String.valueOf(i),
-					file.getString(BT_FILE_NAME), 
-					file.getString(BT_FILE_NAME), 
+					file.getString(BT_FILE_NAME),
+					file.getString(BT_FILE_NAME),
 					null, // Not supported?
-					file.getLong(BT_FILE_SIZE), 
-					file.getLong(BT_FILE_DONE), 
+					file.getLong(BT_FILE_SIZE),
+					file.getLong(BT_FILE_DONE),
 					convertTransmissionPriority(file.getInt(BT_FILE_PRIORITY))));
 				// @formatter:on
 			}
