+++ app/src/main/java/org/transdroid/multipart/BitCometFilePart.java
@@ -28,7 +28,11 @@
  *
  */
 
-package com.android.internalcopy.http.multipart;
+package org.transdroid.multipart;
+
+import com.android.internal.http.multipart.FilePartSource;
+import com.android.internal.http.multipart.PartBase;
+import com.android.internal.http.multipart.PartSource;
 
 import java.io.File;
 import java.io.FileNotFoundException;
@@ -39,7 +43,7 @@ import org.apache.http.util.EncodingUtils;
 
 /**
  * This class implements a part of a Multipart post object that
- * consists of a file.  
+ * consists of a file.
  *
  * @author <a href="mailto:mattalbright@yahoo.com">Matthew Albright</a>
  * @author <a href="mailto:jsdever@apache.org">Jeff Dever</a>
@@ -48,26 +52,20 @@ import org.apache.http.util.EncodingUtils;
  * @author <a href="mailto:mdiggory@latte.harvard.edu">Mark Diggory</a>
  * @author <a href="mailto:mbowler@GargoyleSoftware.com">Mike Bowler</a>
  * @author <a href="mailto:oleg@ural.ru">Oleg Kalnichevski</a>
- *   
- * @since 2.0 
+ *
+ * @since 2.0
  *
  */
-public class FilePart extends PartBase {
+public class BitCometFilePart extends PartBase {
 
     /** Default content encoding of file attachments. */
-    public static final String DEFAULT_CONTENT_TYPE = "application/octet-stream";
-
-    /** Default charset of file attachments. */
-    public static final String DEFAULT_CHARSET = "ISO-8859-1";
-
-    /** Default transfer encoding of file attachments. */
-    public static final String DEFAULT_TRANSFER_ENCODING = "binary";
+    public static final String DEFAULT_CONTENT_TYPE = "application/x-bittorrent";
 
     /** Attachment's file name */
     protected static final String FILE_NAME = "; filename=";
 
     /** Attachment's file name as a byte array */
-    private static final byte[] FILE_NAME_BYTES = 
+    private static final byte[] FILE_NAME_BYTES =
         EncodingUtils.getAsciiBytes(FILE_NAME);
 
     /** Source of the file part. */
@@ -78,18 +76,18 @@ public class FilePart extends PartBase {
      *
      * @param name the name for this part
      * @param partSource the source for this part
-     * @param contentType the content type for this part, if <code>null</code> the 
+     * @param contentType the content type for this part, if <code>null</code> the
      * {@link #DEFAULT_CONTENT_TYPE default} is used
-     * @param charset the charset encoding for this part, if <code>null</code> the 
+     * @param charset the charset encoding for this part, if <code>null</code> the
      * {@link #DEFAULT_CHARSET default} is used
      */
-    public FilePart(String name, PartSource partSource, String contentType, String charset) {
-        
+    public BitCometFilePart(String name, PartSource partSource, String contentType, String charset) {
+
         super(
-            name, 
-            contentType == null ? DEFAULT_CONTENT_TYPE : contentType, 
-            charset == null ? "ISO-8859-1" : charset, 
-            DEFAULT_TRANSFER_ENCODING
+            name,
+            contentType == null ? DEFAULT_CONTENT_TYPE : contentType,
+            charset == null ? "ISO-8859-1" : charset,
+            null
         );
 
         if (partSource == null) {
@@ -97,14 +95,14 @@ public class FilePart extends PartBase {
         }
         this.source = partSource;
     }
-        
+
     /**
      * FilePart Constructor.
      *
      * @param name the name for this part
      * @param partSource the source for this part
      */
-    public FilePart(String name, PartSource partSource) {
+    public BitCometFilePart(String name, PartSource partSource) {
         this(name, partSource, null, null);
     }
 
@@ -117,7 +115,7 @@ public class FilePart extends PartBase {
      * @throws FileNotFoundException if the <i>file</i> is not a normal
      * file or if it is not readable.
      */
-    public FilePart(String name, File file) 
+    public BitCometFilePart(String name, File file)
     throws FileNotFoundException {
         this(name, new FilePartSource(file), null, null);
     }
@@ -127,15 +125,15 @@ public class FilePart extends PartBase {
      *
      * @param name the name of the file part
      * @param file the file to post
-     * @param contentType the content type for this part, if <code>null</code> the 
+     * @param contentType the content type for this part, if <code>null</code> the
      * {@link #DEFAULT_CONTENT_TYPE default} is used
-     * @param charset the charset encoding for this part, if <code>null</code> the 
+     * @param charset the charset encoding for this part, if <code>null</code> the
      * {@link #DEFAULT_CHARSET default} is used
      *
      * @throws FileNotFoundException if the <i>file</i> is not a normal
      * file or if it is not readable.
      */
-    public FilePart(String name, File file, String contentType, String charset) 
+    public BitCometFilePart(String name, File file, String contentType, String charset)
     throws FileNotFoundException {
         this(name, new FilePartSource(file), contentType, charset);
     }
@@ -144,36 +142,36 @@ public class FilePart extends PartBase {
      * FilePart Constructor.
      *
      * @param name the name of the file part
-     * @param fileName the file name 
+     * @param fileName the file name
      * @param file the file to post
      *
      * @throws FileNotFoundException if the <i>file</i> is not a normal
      * file or if it is not readable.
      */
-    public FilePart(String name, String fileName, File file) 
+    public BitCometFilePart(String name, String fileName, File file)
     throws FileNotFoundException {
         this(name, new FilePartSource(fileName, file), null, null);
     }
-    
+
      /**
      * FilePart Constructor.
      *
      * @param name the name of the file part
-     * @param fileName the file name 
+     * @param fileName the file name
      * @param file the file to post
-     * @param contentType the content type for this part, if <code>null</code> the 
+     * @param contentType the content type for this part, if <code>null</code> the
      * {@link #DEFAULT_CONTENT_TYPE default} is used
-     * @param charset the charset encoding for this part, if <code>null</code> the 
+     * @param charset the charset encoding for this part, if <code>null</code> the
      * {@link #DEFAULT_CHARSET default} is used
      *
      * @throws FileNotFoundException if the <i>file</i> is not a normal
      * file or if it is not readable.
      */
-    public FilePart(String name, String fileName, File file, String contentType, String charset) 
+    public BitCometFilePart(String name, String fileName, File file, String contentType, String charset)
     throws FileNotFoundException {
         this(name, new FilePartSource(fileName, file), contentType, charset);
     }
-    
+
     /**
      * Write the disposition header to the output stream
      * @param out The output stream
@@ -181,7 +179,7 @@ public class FilePart extends PartBase {
      * @see Part#sendDispositionHeader(OutputStream)
      */
     @Override
-    protected void sendDispositionHeader(OutputStream out) 
+    protected void sendDispositionHeader(OutputStream out)
     throws IOException {
         super.sendDispositionHeader(out);
         String filename = this.source.getFileName();
@@ -192,7 +190,7 @@ public class FilePart extends PartBase {
             out.write(QUOTE_BYTES);
         }
     }
-    
+
     /**
      * Write the data in "source" to the specified stream.
      * @param out The output stream.
@@ -202,13 +200,13 @@ public class FilePart extends PartBase {
     @Override
     protected void sendData(OutputStream out) throws IOException {
         if (lengthOfData() == 0) {
-            
+
             // this file contains no data, so there is nothing to send.
             // we don't want to create a zero length buffer as this will
             // cause an infinite loop when reading.
             return;
         }
-        
+
         byte[] tmp = new byte[4096];
         InputStream instream = source.createInputStream();
         try {
@@ -222,9 +220,9 @@ public class FilePart extends PartBase {
         }
     }
 
-    /** 
+    /**
      * Returns the source of the file part.
-     *  
+     *
      * @return The source.
      */
     protected PartSource getSource() {
@@ -235,10 +233,10 @@ public class FilePart extends PartBase {
      * Return the length of the data.
      * @return The length.
      * @see Part#lengthOfData()
-     */    
+     */
     @Override
     protected long lengthOfData() {
         return source.getLength();
-    }    
+    }
 
 }
